requirements.txt
streamlit
pytesseract
pillow
openai
fpdf
PyMuPDF
